param module add interface

+++ framework/src/main/java/com/jinhe/tss/framework/web/listener/SessionDestroyedListener.java
@@ -47,9 +47,11 @@ public class SessionDestroyedListener implements HttpSessionListener {
                 "有效期为：" + event.getSession().getMaxInactiveInterval() + " 秒 ");
         
         // 注销在线用户库中对应记录信息，去除登陆用户身份证card信息
-        String token = OnlineUserManagerFactory.getManager().logout(appCode, sessionId);
-        if(token != null) {
-        	Context.destroyIdentityCard(token);
+        if(Context.isOnline()) {
+        	String token = OnlineUserManagerFactory.getManager().logout(appCode, sessionId);
+            if(token != null) {
+            	Context.destroyIdentityCard(token);
+            }
         }
     }
 }
